requirements.txt
customtkinter>=5.1.3
mysql-connector-python>=8.0.33
Pillow>=9.5.0
darkdetect~=0.3.1
typing-extensions~=4.4.0
ttkthemes
cryptography
pyinstaller